refactor: split campaigns page + extracted buttonVariants
