--- src/app/campaigns/isotope.css
@@ -1,34 +1,0 @@
-.campaign-card {
-    transition: transform 0.4s, opacity 0.4s;
-}
-
-.campaign-card {
-    animation: fadeIn 0.5s ease-out;
-}
-
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-        transform: translateY(10px);
-    }
-    to {
-        opacity: 1;
-        transform: translateY(0);
-    }
-}
-
-.campaign-card.isotope-item {
-    z-index: 2;
-}
-
-.campaign-card.isotope-hidden {
-    z-index: 1;
-    pointer-events: none;
-}
-
-/* Animation lors du tri */
-.isotope-item {
-    transition-property: transform, opacity;
-    transition-duration: 0.4s;
-}
-
