+++ src/app/globals.css
@@ -49,20 +49,6 @@ body {
     opacity: 0.01;
 }
 
-.btn-primary {
-    background-color: var(--primary-green);
-    color: var(--white);
-    padding: 0.5rem 1.5rem;
-    border-radius: 4px;
-    transition: all 0.2s ease;
-}
-
-.btn-primary:hover {
-    background-color: var(--dark-green);
-    color: black;
-    border: 1px solid black;
-}
-
