--- src/components/ui/IconButton/IconButton.module.css
@@ -1,23 +1,0 @@
-.button {
-    @apply p-2 rounded-[4px] flex items-center justify-center;
-}
-
-.primary {
-    @apply bg-primary-green text-white;
-
-    &:hover {
-        @apply bg-white text-black border border-black;
-    }
-}
-
-.secondary {
-    @apply bg-beige text-white;
-
-    &:hover {
-        @apply bg-opacity-90;
-    }
-}
-
-.outline {
-    @apply border border-gray-300 text-gray-700;
-}
