+++ src/components/ui/IconButton/IconButton.tsx
@@ -5,28 +5,11 @@ import {ButtonHTMLAttributes} from "react";
 import {cva, VariantProps} from "class-variance-authority";
 
 import {cn} from "@/lib/utils";
-
-import styles from "./IconButton.module.css";
+import {iconButtonVariants} from "@/components/ui/IconButton/IconButtonVariants";
 
 interface IconButtonProps extends ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof iconButtonVariants> {
     icon: LucideIcon;
 }
-
-const iconButtonVariants = cva(
-    styles.button,
-    {
-        variants: {
-            variant: {
-                primary: styles.primary,
-                secondary: styles.secondary,
-                outline: styles.outline
-            }
-        },
-        defaultVariants: {
-            variant: "primary"
-        }
-    }
-);
 
 const IconButton = (props: IconButtonProps) => {
